requirements.txt
fastapi
uvicorn
pydantic
requests
sqlalchemy
pymysql
pandas
mysql-connector-python
websockets
python-dotenv